Implement Age groups page

+++ src/components/CompartmentalModels/ConfigurableParametersForm/children/extraParameters/validators.js
@@ -39,7 +39,7 @@ export const checkErrorsExtraParametersForm = ({extraParameters,handleShowError}
     const {value:valueFirst,errors} = extraParameters[0]
     const {value:valueSecond } = extraParameters[1]         
     const setErrors = extraParameters[0]['custom-actions']['setErrors']
-    if(errors.length == 1 && !errors[0]['type'] && valueFirst!='' && valueSecond!='' && parseInt(valueFirst)<parseInt(valueSecond)){ 
+    if(errors.length == 1 && !errors[0]['type'] && valueFirst!='' && valueSecond!='' && parseFloat(valueFirst)<parseFloat(valueSecond)){ 
       setErrors([])
     }     
   }
